crud_project/requirements.txt
Django==4.2  
djangorestframework==3.14.0  
psycopg2-binary==2.9.6  # PostgreSQL database adapter  
gunicorn==21.2.0  # Production WSGI server  
whitenoise==6.6.0  # Static file handling  
flake8==6.1.0  # Linting for CI/CD  
pytest==7.4.3  # Testing framework  
pytest-django==4.5.2  # Django test integration  
coverage==7.3.2  # Code coverage  
docker==6.1.3  # Docker SDK for Python  
prometheus-client==0.17.1  # Monitoring  
sentry-sdk==1.39.1  # Error tracking  
python-dotenv==1.0.1  # Environment variable management  
pytest-cov
